refactor(deals): the credential's lifecycle reads apart from the roster's
participant_write.go crossed the 500-line cap. Split where the concepts already
divide rather than waived: minting a credential onto a participant, retiring
whatever stood before it, and recording what the relay did move to
credential_lifecycle.go.

The two are governed differently, which is the reason for the seam. A
participant's name and capability are ordinary editable fields. A credential is
a live secret whose one invariant — at most one works at a time — is held by an
index and a supersede that must run in that order, and reading those statements
together without the roster management around them is the point.
